feat: store tokens in system secure storage
This commit also refactors code to implement this feature in a clean way,
and aims to improve existing tests, fix inconsistencies, improve code quality,
and ensure clear separation in related code.
MinecraftAccountManager tests are still not fixed and need to be changed to adapt to the changes,
as we're still evolving the code design approach that we will use. AccountRepository has been introduced,
and its tests are a work in progress. There are a lot of issues that need to be addressed, but generally:

1. Microsoft authentication with both auth and device code needs to be extracted from MinecraftAccountManager into a separate class without depending on AccountRepository or updating/adding the account in either memory or storage.
2. AccountCubit needs a better way of updating the accounts' state and staying in sync with AccountRepository.
3. MicrosoftAccountHandlerCubit needs a better way of communicating with AccountCubit; it should depend on AccountRepository instead and avoid calling AccountCubit, if possible.
4. Maybe we could extract code from MinecraftAccountManager into use case classes and use them in AccountCubit and MicrosoftAccountHandlerCubit directly in a way that makes testing easy?
5. We should update docs/ARCHITECTURE.md with a more general file that includes more details about code style, resources, and guidelines for the project spec.
6. Solve all TODOs added in this branch whenever possible, like completing AccountRepository tests.
7. Quick review to all changes once completed

See https://github.com/KraftLauncher/kraft-launcher/issues/1